Move settingsConfig to forms
