+++ types/applications/forms/settingsConfig.d.ts
